🚚 api: move certification-attestation-pdf to
certfication course src directory
