+++ api/src/certification/course/infrastructure/utils/pdf/files/EN-attestation-template-with-complementary-certification.pdf
