+++ api/src/certification/course/infrastructure/utils/pdf/files/EN-attestation-template-without-complementary-certification.pdf
