+++ api/src/certification/course/infrastructure/utils/pdf/files/FR-attestation-template-with-professionalizing-message-and-without-complementary-certifications.pdf
