+++ api/src/certification/course/infrastructure/utils/pdf/files/FR-attestation-template-without-professionalizing-message-and-with-complementary-certifications.pdf
