+++ api/tests/certification/course/integration/infrastructure/utils/pdf/EN-certification-confirmation-pdf_test.pdf
