+++ api/tests/certification/course/integration/infrastructure/utils/pdf/certification-attestation-pdf_test_full.pdf
