+++ api/tests/certification/course/integration/infrastructure/utils/pdf/stickers/macaron_clea.pdf
