+++ api/tests/certification/course/integration/infrastructure/utils/pdf/stickers/macaron_droit_avance.pdf
