+++ api/tests/certification/course/integration/infrastructure/utils/pdf/stickers/macaron_droit_expert.pdf
