+++ api/tests/certification/course/integration/infrastructure/utils/pdf/stickers/macaron_edu_2nd_initie.pdf
